LayoutTests/crypto/workers/subtle/hmac-import-key-verify.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
    <script>
        worker = startWorker('resources/hmac-import-key-verify.js');
    </script>
    <script src="../../../resources/js-test-post.js"></script>
</body>
</html>
